refactor: split chat reset delivery helper
