+++ ui/src/pages/chat/chat-reset-delivery.ts
@@ -1,0 +1,50 @@
+import type { ChatQueueItem } from "../../lib/chat/chat-types.ts";
+import type { ChatCommandResetOptions } from "./chat-commands.ts";
+import type {
+  ChatOutboxDrainDependencies,
+  QueuedChatSendOptions,
+  QueuedChatSendResult,
+} from "./chat-outbox-drain.ts";
+import { admitQueuedMessageForSession } from "./chat-queue.ts";
+import { cancelChatDelivery } from "./chat-send-composer.ts";
+import type { ChatHost } from "./chat-send-contract.ts";
+import {
+  enqueuePendingSendMessage,
+  reconnectSafeQueuedSendState,
+  setChatError,
+} from "./chat-send-queue-state.ts";
+import { OFFLINE_QUEUE_STORAGE_ERROR } from "./steer-lifecycle.ts";
+
+type DeliverChatQueueItem = (
+  host: ChatHost,
+  item: ChatQueueItem,
+  options?: QueuedChatSendOptions,
+) => Promise<QueuedChatSendResult>;
+
+export function createResetSlashCommandSender(
+  deliverChatQueueItem: DeliverChatQueueItem,
+): ChatOutboxDrainDependencies["sendResetSlashCommand"] {
+  return async (host: ChatHost, message: string, options: ChatCommandResetOptions) => {
+    const item = enqueuePendingSendMessage(
+      host,
+      message,
+      undefined,
+      true,
+      undefined,
+      reconnectSafeQueuedSendState(host),
+    );
+    if (!item || !admitQueuedMessageForSession(host, host.sessionKey, item)) {
+      if (item) {
+        cancelChatDelivery(host, item, { previousDraft: options.previousDraft });
+      }
+      setChatError(host, OFFLINE_QUEUE_STORAGE_ERROR);
+      return;
+    }
+    await deliverChatQueueItem(host, item, {
+      previousDraft: options.previousDraft,
+      restoreDraft: options.restoreDraft,
+      routingSessionKey: host.sessionKey,
+      target: options.target,
+    });
+  };
+}
